JAWT Support JDK >=1.3

+++ demos/MiscDemos/nurbs.java
@@ -102,7 +102,7 @@ public class nurbs extends SimpleGLAnimApplet1
 	};
 
 
-	int theNurb;
+	long theNurb;
 
         public gldemo(int w, int h)
         {
