+++ demos/MiscDemos/tess.java
@@ -132,7 +132,7 @@ public class tess extends Applet {
 	    int i;
 	    glut = new GLUTFuncLightImplWithFonts(gl, glu);
 
-	   int tobj;
+	   long tobj;
 
 	   gl.glClearColor(0.0f, 0.0f, 0.0f, 0.0f);
