+++ demos/MiscDemos/tessdemo.java
@@ -14,8 +14,15 @@
 
 /*
  * $Log$
- * Revision 1.1  2000/11/18 06:53:19  sven
- * Initial revision
+ * Revision 1.2  2001/02/13 05:32:09  sven
+ * JAWT Support JDK >=1.3
+ *
+ * Revision 1.1.1.2  2000/11/18 10:41:27  sven
+ * still working for the initial check in
+ *
+ * Revision 1.1.1.1  2000/11/18 06:53:19  sven
+ * Initial import into CVS.
+ * Version 2.5.0.0 beta 1
  *
  * Revision 1.3f.2.1f  1999/11/16 11:09:09  gareth
  * Added combine callback.  Converted vertices from ints to floats.
@@ -316,7 +323,7 @@ public class tessdemo extends Applet {
 
 	public void tesse( )
 	{
-	   int			tobj;
+	   long			tobj;
 	   double		data[] = new double[3];
 	   int			i, j, point_cnt;
